requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autograd==1.6.2
autograd-gamma==0.5.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bokeh==2.4.3
bottleneck==1.4.0
bs4==0.0.2
cachetools==5.4.0
category-encoders==2.6.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
dask==2.30.0
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
duckdb==1.0.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==0.8.3
faker==19.13.0
fastjsonschema==2.20.0
flask==3.0.3
flask-cors==3.0.10
fonttools==4.53.1
formulaic==1.0.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.0.3
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imagehash==4.3.1
imbalanced-learn==0.12.2
importlib-metadata==7.0.1
importlib-resources==6.4.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==7.8.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==7.4.9
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-server==2.13.0
jupyter-server-terminals==0.5.3
jupyterlab-pygments==0.3.0
jupyterlab-widgets==1.1.9
kiwisolver==1.4.5
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.18
langchain-experimental==0.0.62
langchain-openai==0.1.15
langchain-text-splitters==0.2.2
langsmith==0.1.85
levenshtein==0.12.0
lifelines==0.28.0
llvmlite==0.41.1
locket==1.0.0
lxml==5.2.2
markdown-it-py==3.0.0
markdown2==2.5.0
markdown2docx==0.1.0
markupsafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.1
matplotlib-inline==0.1.7
mdurl==0.1.2
metaphone==0.6
missingno==0.5.2
mistune==0.8.4
multidict==6.0.5
multimethod==1.11.2
mypy-extensions==1.0.0
nbclassic==1.1.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.6.3
notebook==6.5.7
notebook-shim==0.2.4
numba==0.58.1
numpy==1.26.4
openai==1.40.2
openpyxl==3.1.5
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandasai==2.2.14
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.4.0
pip==23.2.1
platformdirs==4.2.2
plotly==5.22.0
ply==3.11
probableparsing==0.0.1
prometheus-client==0.20.0
prompt-toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-stdnum==1.20
pytz==2024.1
pywavelets==1.6.0
pyyaml==6.0.2
pyzmq==26.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
restrictedpython==7.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.12.2
send2trash==1.8.3
setuptools==65.5.0
shap==0.45.1
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlalchemy==2.0.32
sqlglot==25.10.0
sqlglotrs==0.2.8
stack-data==0.5.1
statsmodels==0.14.2
streamlit==1.37.1
streamlit-chat==0.1.1
sweetviz==2.3.1
tableone==0.7.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.5.0
terminado==0.18.1
testpath==0.6.0
threadpoolctl==3.5.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typeguard==4.2.1
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
usaddress==0.5.10
validators==0.22.0
varname==0.8.3
visions==0.7.5
visualimiss==0.0.7
watchdog==4.0.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.3
widgetsnbextension==3.6.8
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.9.4
ydata-profiling==4.7.0
zipp==3.17.0
